requirements.txt
streamlit==1.39.0
langchain-core==0.3.13
langchain-groq==0.1.3
python-dotenv==1.0.1
